// src/types/brief-offer.ts
import {City} from './city.ts';
import {Location} from './location.ts';

export interface BriefOffer {
  id: string;
  title: string;
  type: 'apartment' | 'room' | 'house' | 'hotel';
  price: number;
  city: City;
  location: Location;
  isFavorite: boolean;
  isPremium: boolean;
  rating: number;
  previewImage: string;
}